+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
 #### Exonum core
 
+- `ExecutionError::with_description` method now takes `Into<String>`
+  instead of `String` which allows to pass `&str` directly. (#592)
+
 - POST-requests are now handled with `bodyparser` crate,
   so all the parameters must be passed in the body. (#529)
 
@@ -142,6 +145,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   the `uuid::Uuid`. (#588)
 
 - `Display` trait is implemented for types from the `crypto` module. (#590)
+
+- `transactions!` macro now allows empty body. (#593)
 
 #### exonum-testkit